requirements.txt
google_api_python_client==2.128.0
google_auth_oauthlib==1.2.0
mouse==0.7.1
opencv_python==4.9.0.80
protobuf==5.29.2
PyAutoGUI==0.9.54
pytesseract==0.3.10
pywin32==306
